+++ src/wrapping.rs
@@ -503,6 +503,16 @@ impl<F: Fixed> Wrapping<F> {
         self.0.trailing_zeros()
     }
 
+    /// Integer base-2 logarithm, rounded down.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the fixed-point number is ≤ 0.
+    #[inline]
+    pub fn int_log2(self) -> i32 {
+        self.0.int_log2()
+    }
+
     /// Shifts to the left by `n` bits, wrapping the truncated bits to the right end.
     ///
     /// # Examples